Get, edit, delete and print file are all working with directory server. Need to do create and list

+++ Directory Service (part 3 of 4)/clientLibrary.py
@@ -16,9 +16,18 @@ def listFiles(ip, port):
 # passed on the fileserver with provided port and ip
 # If the file does not exist, returns -1
 def getFile(ip, port, filename):
+    # Directory service access
     location = 'http://{}:{}/filedir/{}'.format(ip, port, filename)
     r = requests.get(location)
-    json_data = json.loads(r.text)  # JSON to dict (JSON
+    direct_data = json.loads(r.text)  # JSON to dict (JSON)
+    if direct_data['ip'] == -1:
+        print("(getFile) File does not exist on server")
+        return -1
+
+    # File Server access
+    location = 'http://{}:{}/filedir/{}'.format(direct_data['ip'], direct_data['port'], direct_data['filename'])
+    r = requests.get(location)
+    json_data = json.loads(r.text)  # JSON to dict (JSON)
     if 'success' in json_data:
         if json_data['success'] == False:
             print("(getFile) File does not exist on server")
@@ -28,10 +37,19 @@ def getFile(ip, port, filename):
 
 # Takes a dictionary of the file, replaces the data with newText and
 # PUTs it on to the file server
-# Returns -1 if file is behind on version
+# Returns -1 if file is behind on version or if it doesn't exist on the server
 def editFile(ip, port, fileDict, newText):
-    fileDict['data'] = newText
+    # Directory service access
     location = 'http://{}:{}/filedir/{}'.format(ip, port, fileDict['filename'])
+    r = requests.get(location)
+    direct_data = json.loads(r.text)  # JSON to dict (JSON)
+    if direct_data['ip'] == -1:
+        print("(editFile) File does not exist on server")
+        return -1
+
+    # Edit the file
+    fileDict['data'] = newText
+    location = 'http://{}:{}/filedir/{}'.format(direct_data['ip'], direct_data['port'], direct_data['filename'])
     r = requests.put(location, json={'version': fileDict['version'], 'data':fileDict['data']})
     json_data = json.loads(r.text)
     if json_data['success'] == 'notOnServer':
@@ -42,6 +60,9 @@ def editFile(ip, port, fileDict, newText):
 
 # Creates a new file called filename with content data (POST)
 def createFile(ip, port, filename, data):
+    # Get info for a single file server
+
+    # Create the file
     location = 'http://{}:{}/filedir'.format(ip, port)
     r = requests.post(location, json={'filename': filename, 'version': 0, 'data': data})
     json_data = json.loads(r.text)
@@ -50,7 +71,16 @@ def createFile(ip, port, filename, data):
 
 # Deletes the file on the server, returns -1 if file not found
 def deleteFile(ip, port, filename):
+    # Directory service access
     location = 'http://{}:{}/filedir/{}'.format(ip, port, filename)
+    r = requests.get(location)
+    direct_data = json.loads(r.text)  # JSON to dict (JSON)
+    if direct_data['ip'] == -1:
+        print("(getFile) File does not exist on server")
+        return -1
+
+    # Delete the file
+    location = 'http://{}:{}/filedir/{}'.format(direct_data['ip'], direct_data['port'], direct_data['filename'])
     r = requests.delete(location)
     json_data = json.loads(r.text)  # JSON to dict (JSON
     if 'success' in json_data:
@@ -60,6 +90,7 @@ def deleteFile(ip, port, filename):
         elif json_data['success'] == True:
             print("Successful deletion")
 
+# Takes a dictionary of file data and prints it
 def printFile(fileDict):
 
     print("--------------------------")
